+++ tests/src/sdmp/test_netif.c
@@ -46,7 +46,7 @@ static vs_netif_rx_cb_t callback_rx_cb;
 
 /**********************************************************/
 static int
-test_netif_tx(const uint8_t *data, const size_t data_sz) {
+test_netif_tx(const uint8_t *data, const uint16_t data_sz) {
     int ret_code;
 
     is_client_call = !is_client_call;
